IN-38 refactor: change file structure
